move dataclasses and bq schemas into their own file
