+++ pipeline/metadata/flatten_base.py
@@ -3,157 +3,11 @@
 from __future__ import absolute_import
 from __future__ import annotations  # required to use class as a type inside the class
 
-import dataclasses
-from dataclasses import dataclass
 import os
 from typing import Optional, List, Dict, Any, Union
 
 from pipeline.metadata.blockpage import BlockpageMatcher
-
-# pylint: disable=too-many-instance-attributes
-
-
-@dataclass
-class SatelliteAnswer():
-  """Class for keeping track of Satellite answer content"""
-  # Keys
-  ip: str
-  # Metadata
-  asnum: Optional[int] = None
-  asname: Optional[str] = None
-  http: Optional[str] = None
-  cert: Optional[str] = None
-  matches_control: Optional[str] = None
-
-
-@dataclass
-class SatelliteAnswerMetadata(SatelliteAnswer):
-  """Satellite Answer Metadata.
-
-  When this metadata is being passed around
-  it needs an additional date field to keep track of when it's valid.
-  In the final SatelliteAnswer object written to bigquery
-  we don't include that field.
-  """
-  date: str = ''
-
-
-@dataclass
-class IpMetadata():
-  """Class for keeping track of ip metadata"""
-  # Keys
-  ip: str
-  date: str
-  # Metadata
-  netblock: Optional[str] = None
-  asn: Optional[int] = None
-  as_name: Optional[str] = None
-  as_full_name: Optional[str] = None
-  as_class: Optional[str] = None
-  country: Optional[str] = None
-  organization: Optional[str] = None
-  # Satellite Metadata
-  name: Optional[str] = None
-
-
-# Custom Type
-# All or part of a scan row to be written to bigquery
-# ex (only scan data): {'domain': 'test.com', 'ip': '1.2.3.4', 'success' true}
-# ex (only ip metadata): {'asn': 13335, 'as_name': 'CLOUDFLAREINC'}
-# ex (both): {'domain': 'test.com', 'ip': '1.2.3.4', 'asn': 13335}
-@dataclass
-class Row:  # Corrosponds to BASE_BIGQUERY_SCHEMA
-  """Class for keeping track of row content"""
-  domain: Optional[str] = None
-  category: Optional[str] = None
-  ip: Optional[str] = None
-  date: Optional[str] = None
-  start_time: Optional[str] = None
-  end_time: Optional[str] = None
-  error: Optional[str] = None
-  anomaly: Optional[bool] = None
-  success: Optional[bool] = None
-  is_control: Optional[bool] = None
-  controls_failed: Optional[bool] = None
-  measurement_id: Optional[str] = None
-  source: Optional[str] = None
-
-  # Metadata
-  netblock: Optional[str] = None
-  asn: Optional[int] = None
-  as_name: Optional[str] = None
-  as_full_name: Optional[str] = None
-  as_class: Optional[str] = None
-  country: Optional[str] = None
-  organization: Optional[str] = None
-
-  # imitation update method that matches the semantics of python's dict update
-  def update(self, new: Row) -> None:
-    for field in dataclasses.fields(new):
-      value = getattr(new, field.name)
-      if value is not None and value != []:
-        setattr(self, field.name, value)
-
-
-@dataclass
-class HyperquackRow(Row):
-  """Class for hyperquack specific fields"""
-  blockpage: Optional[bool] = None
-  page_signature: Optional[str] = None
-  stateful_block: Optional[bool] = None
-  received_status: Optional[str] = None
-  received_body: Optional[str] = None
-  received_tls_version: Optional[int] = None
-  received_tls_cipher_suite: Optional[int] = None
-  received_tls_cert: Optional[str] = None
-  received_headers: List[str] = dataclasses.field(default_factory=list)
-
-
-@dataclass
-class SatelliteRow(Row):
-  """Class for satellite specific fields"""
-  rcode: Optional[int] = None
-  name: Optional[str] = None
-  is_control_ip: Optional[bool] = None
-  average_confidence: Optional[float] = None
-  untagged_controls: Optional[bool] = None
-  untagged_response: Optional[bool] = None
-  excluded: Optional[bool] = None
-  exclude_reason: Optional[str] = None
-  has_type_a: Optional[bool] = None
-  received: List[SatelliteAnswer] = dataclasses.field(default_factory=list)
-  matches_confidence: List[float] = dataclasses.field(default_factory=list)
-
-
-@dataclass
-class BlockpageRow():
-  """Class for blockpage specific fields"""
-  domain: Optional[str] = None
-  ip: Optional[str] = None
-  date: Optional[str] = None
-  start_time: Optional[str] = None
-  end_time: Optional[str] = None
-  success: Optional[bool] = None
-  https: Optional[bool] = None
-  source: Optional[str] = None
-
-  blockpage: Optional[bool] = None
-  page_signature: Optional[str] = None
-
-  received_status: Optional[str] = None
-  received_body: Optional[str] = None
-  received_tls_version: Optional[int] = None
-  received_tls_cipher_suite: Optional[int] = None
-  received_tls_cert: Optional[str] = None
-  received_headers: List[str] = dataclasses.field(default_factory=list)
-
-  # imitation update method that matches the semantics of python's dict update
-  def update(self, new: Union[BlockpageRow, HyperquackRow]) -> None:
-    for field in dataclasses.fields(new):
-      value = getattr(new, field.name)
-      if value is not None and value != []:
-        setattr(self, field.name, value)
-
+from pipeline.metadata.schema import HyperquackRow
 
 # pylint: enable=too-many-instance-attributes
