+++ pipeline/metadata/flatten_hyperquack.py
@@ -6,7 +6,7 @@ import re
 from typing import Optional, Any, Iterator
 
 from pipeline.metadata import flatten_base
-from pipeline.metadata.flatten_base import HyperquackRow
+from pipeline.metadata.schema import HyperquackRow
 from pipeline.metadata.blockpage import BlockpageMatcher
 from pipeline.metadata.domain_categories import DomainCategoryMatcher
